backend/requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.22.0
psycopg2-binary>=2.9.0
python-dotenv>=1.0.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
bcrypt>=4.1.0
slowapi>=0.1.9
pytest>=8.0.0
pytesseract>=0.3.10
Pillow>=10.0.0
openai>=1.0.0
boto3>=1.28.0
opencv-python-headless>=4.8.0
numpy>=1.24.0
playwright>=1.40.0
xhtml2pdf>=0.2.13
pymupdf>=1.24.0
python-docx>=1.1.0
docx2pdf>=0.1.8
